redirect so users can't apply prematurely
